deleted notused files and code
