+++ helpers/check_new_parcels.js
@@ -1,6 +1,6 @@
 const {HTTPyard} = require('../http/yard');
 const ParcelModel = require('../db/schema/parcels');
-const {findMissingNumbers} = require('./arrayUtils');
+const {findMissingNumbers} = require('./array_utils');
 const telegram = require('./telegram');
 const ChatIdModel = require('../db/schema/chatId');
