+++ package.json
@@ -13,7 +13,6 @@
   "dependencies": {
     "axios": "^1.6.8",
     "dotenv": "^16.0.3",
-    "express": "^4.18.2",
     "node-cron": "^3.0.3",
     "node-telegram-bot-api": "^0.65.1",
     "nodemon": "^3.1.0",
